HHH-10313 Make SessionImplementor extend WrapperOptions
As per the TODO in AbstractStandardBasicType. This removes the need to allocate a new WrapperOptions for every call and saves a lot of allocations.
